+++ platform/src/common/components/Map/AirQoMap.jsx
@@ -1,64 +1,44 @@
-import React, { useEffect, useRef, useState, useCallback } from 'react';
+// AirQoMap.jsx
+import React, { useEffect, useRef, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import mapboxgl from 'mapbox-gl';
 import axios from 'axios';
-import Supercluster from 'supercluster';
 import { useWindowSize } from '@/lib/windowSize';
-import { getMapReadings } from '@/core/apis/DeviceRegistry';
 import {
-  setOpenLocationDetails,
-  setSelectedLocation,
   setMapLoading,
-  setCenter,
-  setZoom,
-  setSelectedNode,
   reSetMap,
-  setSelectedWeeklyPrediction,
-  setMapReadingsData,
-  setWaqData,
   clearData,
 } from '@/lib/store/services/map/MapSlice';
-import {
-  CustomGeolocateControl,
-  CustomZoomControl,
-  IconButton,
-  LoadingOverlay,
-} from './components/MapControls';
 import LayerModal from './components/LayerModal';
 import Loader from '@/components/Spinner';
 import Toast from '../Toast';
-import { AQI_FOR_CITIES } from './data/Cities';
 import { AirQualityLegend } from './components/Legend';
-import {
-  createPopupHTML,
-  createClusterNode,
-  UnclusteredNode,
-  getAQICategory,
-  images,
-} from './components/MapNodes';
 import { mapStyles, mapDetails } from './data/constants';
 import LayerIcon from '@/icons/map/layerIcon';
 import RefreshIcon from '@/icons/map/refreshIcon';
 import ShareIcon from '@/icons/map/shareIcon';
 import PropTypes from 'prop-types';
+import {
+  useMapData,
+  useRefreshMap,
+  useShareLocation,
+  CustomZoomControl,
+  CustomGeolocateControl,
+  IconButton,
+  LoadingOverlay,
+} from './functions';
 
 const AirQoMap = ({ customStyle, mapboxApiAccessToken, pollutant }) => {
   const dispatch = useDispatch();
   const mapContainerRef = useRef(null);
   const { width } = useWindowSize();
-  const markersRef = useRef([]);
-  const mapRef = useRef();
-  const indexRef = useRef();
   const [isOpen, setIsOpen] = useState(false);
-  const [refresh, setRefresh] = useState(false);
   const [loading, setLoading] = useState(false);
   const [loadingOthers, setLoadingOthers] = useState(false);
   const urls = new URL(window.location.href);
   const urlParams = new URLSearchParams(urls.search);
   const mapData = useSelector((state) => state.map);
   const selectedNode = useSelector((state) => state.map.selectedNode);
-  const mapReadingsData = useSelector((state) => state.map.mapReadingsData);
-  const waqData = useSelector((state) => state.map.waqData);
   const [toastMessage, setToastMessage] = useState({
     message: '',
     type: '',
@@ -84,20 +64,22 @@ const AirQoMap = ({ customStyle, mapboxApiAccessToken, pollutant }) => {
 
   // Stop loaders after 10 seconds
   useEffect(() => {
-    const loaderTimer = setTimeout(() => {
-      setLoading(false);
-    }, 10000);
-
-    return () => clearTimeout(loaderTimer);
+    if (loading) {
+      const loaderTimer = setTimeout(() => {
+        setLoading(false);
+      }, 10000);
+      return () => clearTimeout(loaderTimer);
+    }
   }, [loading]);
 
-  // Stop map loading skeleton after 2 seconds
+  // Stop map loading skeleton after 2 seconds when a node is selected
   useEffect(() => {
-    const skeletonTimer = setTimeout(() => {
-      dispatch(setMapLoading(false));
-    }, 2000);
-
-    return () => clearTimeout(skeletonTimer);
+    if (selectedNode) {
+      const skeletonTimer = setTimeout(() => {
+        dispatch(setMapLoading(false));
+      }, 2000);
+      return () => clearTimeout(skeletonTimer);
+    }
   }, [dispatch, selectedNode]);
 
   // Clear map state if lat, lng, or zm params are missing
@@ -107,19 +89,42 @@ const AirQoMap = ({ customStyle, mapboxApiAccessToken, pollutant }) => {
     }
   }, [lat, lng, zm, dispatch]);
 
+  // Custom hook to fetch data and manage map data
+  const { mapRef, fetchAndProcessData, clusterUpdate } = useMapData({
+    NodeType,
+    selectedNode,
+    mapStyle,
+    pollutant,
+    refresh: null, // Pass any refresh dependencies if needed
+    setLoading,
+    setLoadingOthers,
+  });
+
+  // Use custom hooks for refreshing map and sharing location
+  const refreshMap = useRefreshMap(
+    setToastMessage,
+    mapRef,
+    dispatch,
+    selectedNode,
+  );
+  const shareLocation = useShareLocation(setToastMessage, mapRef);
+
   // Initialize the map and add controls
   useEffect(() => {
     const initializeMap = async () => {
       try {
         mapboxgl.accessToken = mapboxApiAccessToken;
+        const initialCenter = [
+          lng ? parseFloat(lng) : mapData.center.longitude,
+          lat ? parseFloat(lat) : mapData.center.latitude,
+        ];
+        const initialZoom = zm ? parseFloat(zm) : mapData.zoom;
+
         const map = new mapboxgl.Map({
           container: mapContainerRef.current,
           style: mapStyle,
-          center: [
-            lng || mapData.center.longitude,
-            lat || mapData.center.latitude,
-          ],
-          zoom: zm || mapData.zoom,
+          center: initialCenter,
+          zoom: initialZoom,
         });
 
         mapRef.current = map;
@@ -137,6 +142,9 @@ const AirQoMap = ({ customStyle, mapboxApiAccessToken, pollutant }) => {
             map.addControl(zoomControl, 'bottom-right');
             map.addControl(geolocateControl, 'bottom-right');
           }
+
+          // Fetch and process data once the map is loaded
+          fetchAndProcessData();
         });
       } catch (error) {
         console.error('Error initializing the Map: ', error);
@@ -151,7 +159,20 @@ const AirQoMap = ({ customStyle, mapboxApiAccessToken, pollutant }) => {
         mapRef.current.remove();
       }
     };
-  }, [mapStyle, NodeType, mapboxApiAccessToken, width, selectedNode]);
+  }, [
+    mapStyle,
+    NodeType,
+    mapboxApiAccessToken,
+    width,
+    selectedNode,
+    fetchAndProcessData,
+    lat,
+    lng,
+    zm,
+    mapData.center.latitude,
+    mapData.center.longitude,
+    mapData.zoom,
+  ]);
 
   // Fly to new center and zoom when mapData changes
   useEffect(() => {
@@ -165,342 +186,13 @@ const AirQoMap = ({ customStyle, mapboxApiAccessToken, pollutant }) => {
         });
       }
     }
-  }, [mapData.center, mapData.zoom, mapRef.current]);
+  }, [mapData.center, mapData.zoom]);
 
-  /**
-   * Handling data section
-   */
-  const createFeature = useCallback(
-    (id, name, coordinates, aqi, no2, pm10, pm2_5, time, forecast) => ({
-      type: 'Feature',
-      geometry: { type: 'Point', coordinates },
-      properties: {
-        _id: id,
-        location: name,
-        airQuality: aqi?.category,
-        no2,
-        pm10,
-        pm2_5,
-        createdAt: time,
-        time,
-        aqi: aqi || 'undefined',
-        forecast,
-      },
-    }),
-    [],
-  );
-
-  // Forecast data
-  const getForecastForPollutant = useCallback(
-    (cityData) => {
-      if (!cityData || !cityData.forecast?.daily) return null;
-
-      const dailyForecast = cityData.forecast.daily;
-      const forecastDataArray =
-        pollutant === 'pm2_5' ? dailyForecast.pm25 : dailyForecast.pm10;
-
-      return forecastDataArray?.map((forecastData) => ({
-        [pollutant]: forecastData.avg,
-        time: forecastData.day,
-      }));
-    },
-    [pollutant],
-  );
-
-  // formatting waq data
-  const fetchAndProcessWaqData = useCallback(
-    async (cities) => {
-      setLoadingOthers(true);
-      try {
-        const responses = await Promise.allSettled(
-          cities.map((city) => axios.get(`/api/proxy?city=${city}`)),
-        );
-
-        // Filter and process the fulfilled responses
-        const fulfilledResponses = responses.filter(
-          (response) =>
-            response.status === 'fulfilled' && response.value?.data?.data?.city,
-        );
-
-        // Stop loading as soon as we process the fulfilled responses
-        if (fulfilledResponses.length > 0) {
-          setLoadingOthers(false);
-        }
-
-        // Map the data into the desired format
-        return fulfilledResponses
-          .map((response) => {
-            const cityData = response.value.data.data;
-            const waqiPollutant = pollutant === 'pm2_5' ? 'pm25' : pollutant;
-            const aqi = getAQICategory(
-              pollutant,
-              cityData.iaqi[waqiPollutant]?.v,
-            );
-
-            return createFeature(
-              cityData.idx,
-              cityData.city.name,
-              [cityData.city.geo[1], cityData.city.geo[0]],
-              aqi,
-              cityData.iaqi.no2?.v,
-              cityData.iaqi.pm10?.v,
-              cityData.iaqi.pm25?.v,
-              cityData.time.iso,
-              getForecastForPollutant(cityData),
-            );
-          })
-          .filter(Boolean);
-      } catch (error) {
-        console.error('Error fetching AQI data:', error);
-        return [];
-      } finally {
-        setLoadingOthers(false); // Ensures loading is stopped in case of error
-      }
-    },
-    [pollutant, createFeature, getForecastForPollutant],
-  );
-
-  // formatting readings data
-  const fetchAndProcessMapReadings = useCallback(async () => {
-    try {
-      setLoading(true);
-
-      const response = await getMapReadings();
-      console.info('response', response);
-
-      if (response.success && response?.measurements?.length > 0) {
-        // Process and filter the data in the same step
-        return response.measurements
-          .filter(
-            (item) => item.siteDetails && item.no2 && item.pm10 && item.pm2_5,
-          )
-          .map((item) => {
-            const aqi = getAQICategory(pollutant, item[pollutant]?.value);
-            return createFeature(
-              item.siteDetails._id,
-              item.siteDetails.name,
-              [
-                item.siteDetails.approximate_longitude,
-                item.siteDetails.approximate_latitude,
-              ],
-              aqi,
-              item.no2?.value,
-              item.pm10?.value,
-              item.pm2_5?.value,
-              item.time,
-              null,
-            );
-          });
-      }
-
-      setLoading(false);
-      return [];
-    } catch (error) {
-      console.error('Error fetching map readings data:', error);
-      return [];
-    } finally {
-      setLoading(false);
-    }
-  }, [pollutant, createFeature]);
-
-  const fetchAndProcessData = useCallback(async () => {
-    if (!mapReadingsData.length) {
-      const newMapReadingsData = await fetchAndProcessMapReadings();
-      dispatch(setMapReadingsData(newMapReadingsData));
-    }
-
-    if (!waqData.length) {
-      const newWaqData = await fetchAndProcessWaqData(AQI_FOR_CITIES);
-      dispatch(setWaqData(newWaqData));
-    }
-  }, [
-    dispatch,
-    fetchAndProcessMapReadings,
-    fetchAndProcessWaqData,
-    mapReadingsData,
-    waqData,
-  ]);
-
-  const getTwoMostCommonAQIs = useCallback((cluster) => {
-    const leaves = indexRef.current.getLeaves(
-      cluster.properties.cluster_id,
-      Infinity,
-    );
-    const aqiCounts = leaves.reduce((acc, leaf) => {
-      const aqi = leaf.properties.airQuality;
-      acc[aqi] = (acc[aqi] || 0) + 1;
-      return acc;
-    }, {});
-
-    const sortedAQIs = Object.entries(aqiCounts).sort((a, b) => b[1] - a[1]);
-    const mostCommonAQIs =
-      sortedAQIs.length > 1
-        ? sortedAQIs.slice(0, 2)
-        : [sortedAQIs[0], sortedAQIs[0]];
-
-    return leaves
-      .filter((leaf) =>
-        mostCommonAQIs
-          .map((aqi) => aqi[0])
-          .includes(leaf.properties.airQuality),
-      )
-      .map((leaf) => ({
-        aqi: leaf.properties.aqi,
-        pm2_5: leaf.properties.pm2_5,
-        pm10: leaf.properties.pm10,
-        no2: leaf.properties.no2,
-      }));
-  }, []);
-
-  const clusterUpdate = useCallback(async () => {
-    const map = mapRef.current;
-
-    const index = new Supercluster({
-      radius: NodeType === 'Emoji' ? 40 : NodeType === 'Node' ? 60 : 80,
-    });
-
-    indexRef.current = index;
-
-    const handleClusterUpdate = () => {
-      try {
-        const zoom = map.getZoom();
-        const bounds = map.getBounds();
-        const bbox = [
-          bounds.getWest(),
-          bounds.getSouth(),
-          bounds.getEast(),
-          bounds.getNorth(),
-        ];
-
-        const clusters = index.getClusters(bbox, Math.floor(zoom));
-
-        markersRef.current.forEach((marker) => marker.remove());
-        markersRef.current = [];
-
-        clusters.forEach((feature) => {
-          const el = document.createElement('div');
-          el.style.cursor = 'pointer';
-
-          if (!feature.properties.cluster) {
-            el.style.zIndex = 1;
-            el.innerHTML = UnclusteredNode({ feature, NodeType, selectedNode });
-
-            const popup = new mapboxgl.Popup({
-              offset:
-                NodeType === 'Node' ? 35 : NodeType === 'Number' ? 42 : 58,
-              closeButton: false,
-              maxWidth: 'none',
-              className: 'my-custom-popup hidden md:block',
-            }).setHTML(createPopupHTML({ feature, images }));
-
-            const marker = new mapboxgl.Marker(el)
-              .setLngLat(feature.geometry.coordinates)
-              .setPopup(popup)
-              .addTo(map);
-
-            el.addEventListener('mouseenter', () => {
-              marker.togglePopup();
-              el.style.zIndex = 9999;
-            });
-            el.addEventListener('mouseleave', () => {
-              marker.togglePopup();
-              el.style.zIndex = 1;
-            });
-
-            el.addEventListener('click', () => {
-              if (selectedNode === feature.properties._id) return;
-
-              dispatch(setSelectedNode(feature.properties._id));
-              dispatch(setSelectedWeeklyPrediction(null));
-              dispatch(setMapLoading(true));
-              dispatch(setOpenLocationDetails(true));
-              dispatch(setSelectedLocation(feature.properties));
-
-              dispatch(
-                setCenter({
-                  latitude: feature.geometry.coordinates[1],
-                  longitude: feature.geometry.coordinates[0],
-                }),
-              );
-              dispatch(setZoom(15));
-            });
-
-            markersRef.current.push(marker);
-          } else {
-            el.zIndex = 444;
-            el.className =
-              'clustered flex justify-center items-center bg-white rounded-full p-2 shadow-md';
-
-            const mostCommonAQIs = getTwoMostCommonAQIs(feature);
-            feature.properties.aqi = mostCommonAQIs;
-
-            el.innerHTML = createClusterNode({ feature, NodeType });
-            const marker = new mapboxgl.Marker(el)
-              .setLngLat(feature.geometry.coordinates)
-              .addTo(map);
-
-            el.addEventListener('click', () => {
-              map.flyTo({
-                center: feature.geometry.coordinates,
-                zoom: zoom + 2,
-              });
-            });
-
-            markersRef.current.push(marker);
-          }
-        });
-      } catch (error) {
-        console.error('Error updating clusters: ', error);
-      }
-    };
-
-    map.on('zoomend', handleClusterUpdate);
-    map.on('moveend', handleClusterUpdate);
-
-    if (mapReadingsData?.length > 0) {
-      try {
-        index.load(mapReadingsData);
-        handleClusterUpdate();
-      } catch (error) {
-        console.error(
-          'Error loading map readings data into Supercluster: ',
-          error,
-        );
-      }
-    }
-
-    if (waqData?.length > 0) {
-      try {
-        const data = [...mapReadingsData, ...waqData];
-        index.load(data);
-        handleClusterUpdate();
-      } catch (error) {
-        console.error('Error loading AQI data into Supercluster: ', error);
-      }
-    }
-
-    return () => {
-      map.off('zoomend', handleClusterUpdate);
-      map.off('moveend', handleClusterUpdate);
-    };
-  }, [
-    selectedNode,
-    NodeType,
-    mapStyle,
-    pollutant,
-    refresh,
-    waqData,
-    mapReadingsData,
-    width,
-    getTwoMostCommonAQIs,
-  ]);
-
+  // Update clusters whenever data changes
   useEffect(() => {
-    fetchAndProcessData();
-  }, [fetchAndProcessData]);
-
-  useEffect(() => {
-    clusterUpdate();
+    if (mapRef.current) {
+      clusterUpdate();
+    }
   }, [clusterUpdate]);
 
   /**
@@ -513,6 +205,7 @@ const AirQoMap = ({ customStyle, mapboxApiAccessToken, pollutant }) => {
 
       if (!map) return;
 
+      // Remove existing boundaries if any
       if (map.getLayer('location-boundaries')) {
         map.removeLayer('location-boundaries');
       }
@@ -559,9 +252,9 @@ const AirQoMap = ({ customStyle, mapboxApiAccessToken, pollutant }) => {
             },
           });
 
-          const { lat, lon } = data[0];
+          const { lat: boundaryLat, lon: boundaryLon } = data[0];
           map.flyTo({
-            center: [lon, lat],
+            center: [parseFloat(boundaryLon), parseFloat(boundaryLat)],
             zoom: mapData.location.city && mapData.location.country ? 10 : 5,
           });
 
@@ -586,73 +279,24 @@ const AirQoMap = ({ customStyle, mapboxApiAccessToken, pollutant }) => {
   }, [mapData.location]);
 
   /**
-   * Refresh the map
-   */
-  const refreshMap = useCallback(() => {
-    const map = mapRef.current;
-
-    // Ensure the map has a valid style URL to reset it
-    const originalStyle =
-      map.getStyle().sprite.split('/').slice(0, -1).join('/') + '/style.json';
-
-    map.setStyle(originalStyle);
-
-    setRefresh((prev) => !prev);
-    if (selectedNode) {
-      dispatch(setSelectedNode(null));
-    }
-    dispatch(reSetMap());
-
-    setToastMessage({
-      message: 'Map refreshed successfully',
-      type: 'success',
-      bgColor: 'bg-blue-600',
-    });
-  }, [dispatch, selectedNode]);
-
-  /**
-   * Share location URL
-   * @sideEffect
-   * - Copy URL to clipboard
-   * - Display toast notification
-   * @returns {void}
-   */
-  const shareLocation = useCallback(() => {
-    try {
-      const map = mapRef.current;
-      const center = map.getCenter();
-      const zoom = map.getZoom();
-      const currentUrl = window.location.href;
-
-      const url = new URL(currentUrl);
-      url.searchParams.set('lat', center.lat.toFixed(4));
-      url.searchParams.set('lng', center.lng.toFixed(4));
-      url.searchParams.set('zm', zoom.toFixed(2));
-
-      navigator.clipboard.writeText(url.toString());
-
-      setToastMessage({
-        message: 'Location URL copied to clipboard',
-        type: 'success',
-        bgColor: 'bg-blue-600',
-      });
-    } catch (error) {
-      console.error('Error sharing location:', error);
-      setToastMessage({
-        message: 'Failed to copy location URL to clipboard',
-        type: 'error',
-      });
-    }
-  }, []);
-
-  /**
-   * Resize the map
-   * @sideEffect
-   * - Resizes the map when the window is resized and side bar is closed
+   * Resize the map on window resize or when a node is selected/deselected
    */
   useEffect(() => {
-    mapRef.current.resize();
-  }, [window.innerWidth, window.innerHeight, selectedNode]);
+    const handleResize = () => {
+      if (mapRef.current) {
+        mapRef.current.resize();
+      }
+    };
+
+    window.addEventListener('resize', handleResize);
+
+    // Trigger resize when selectedNode changes
+    handleResize();
+
+    return () => {
+      window.removeEventListener('resize', handleResize);
+    };
+  }, [selectedNode]);
 
   return (
     <div className="relative w-full h-full">
@@ -660,16 +304,16 @@ const AirQoMap = ({ customStyle, mapboxApiAccessToken, pollutant }) => {
       <div ref={mapContainerRef} className={customStyle} />
 
       {/* Air Quality Legend */}
-      {width >= 1024 || !selectedNode ? (
+      {(width >= 1024 || !selectedNode) && (
         <div className="relative left-4 z-50 md:block">
           <div className="absolute bottom-2 z-[900]">
             <AirQualityLegend pollutant={pollutant} />
           </div>
         </div>
-      ) : null}
+      )}
 
       {/* Map Controls */}
-      {width >= 1024 || !selectedNode ? (
+      {(width >= 1024 || !selectedNode) && (
         <div className="absolute top-4 right-0 z-40 flex flex-col gap-4">
           <IconButton
             onClick={() => setIsOpen(true)}
@@ -687,7 +331,7 @@ const AirQoMap = ({ customStyle, mapboxApiAccessToken, pollutant }) => {
             icon={<ShareIcon />}
           />
         </div>
-      ) : null}
+      )}
 
       {/* Loading Overlay */}
       {loading && (
@@ -731,9 +375,9 @@ const AirQoMap = ({ customStyle, mapboxApiAccessToken, pollutant }) => {
 };
 
 AirQoMap.propTypes = {
-  customStyle: PropTypes.object,
-  mapboxApiAccessToken: PropTypes.string,
-  pollutant: PropTypes.string,
+  customStyle: PropTypes.string, // Assuming customStyle is a string of class names
+  mapboxApiAccessToken: PropTypes.string.isRequired,
+  pollutant: PropTypes.string.isRequired,
 };
 
 export default React.memo(AirQoMap);
