+++ platform/src/common/components/Map/functions/index.jsx
@@ -1,0 +1,9 @@
+export { default as useMapData } from './fetchData';
+export {
+  useRefreshMap,
+  useShareLocation,
+  CustomZoomControl,
+  CustomGeolocateControl,
+  IconButton,
+  LoadingOverlay,
+} from './mapControls';
